Update gsutil command in README
I tested this tutorial today. The checkpoint conversion script expects the .pth file to be in the root of the bucket so that when the contents are copied to the user's system, the .pth file is in /tmp/<bucket_name>. With the previous gsutil cp command, the .pth file ended up in /tmp/<bucket_name>/llama-2-7b. This change only copies the file in the llama-2-7b directory rather than the directory itself. I tested this tutorial with the change and it worked.